- Added staff charge restriction per project
This update introduces a new attribute in the project model that specifies if staff charges are allowed for the project. Functionality has been added to check this attribute before creating a staff charge, and to limit displayed projects in the staff charges template to those allowing staff charges. A test has also been created to verify these changes.